requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
gunicorn==20.1.0
huffman-tree==0.0.1
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
Werkzeug==2.1.2
zipp==3.8.0
